refactor: zkvm details logic
