+++ lib/constants.ts
@@ -1,3 +1,5 @@
+import type { MetricThresholds } from "./types"
+
 export const SITE_NAME = "Ethproofs"
 export const SITE_DESCRIPTION = "Building a fully SNARKed Ethereum"
 export const SITE_URL = process.env.SITE_URL || "https://ethproofs.org"
@@ -37,3 +39,26 @@ export const CHART_RANGES = [7, 30, 90] as const
 
 export const BENCHMARK_LOWER_THRESHOLD = 45_000 // milliseconds
 export const BENCHMARK_UPPER_THRESHOLD = 90_000 // milliseconds
+
+export const ZKVM_THRESHOLDS = {
+  size_bytes: {
+    red: 512 * 1024, // 512 KiB
+    yellow: 32 * 1024, // 32 KiB
+    green: 0,
+  },
+  verification_ms: {
+    red: 16, // 16ms
+    yellow: 1, // 1ms
+    green: 0,
+  },
+  security_target_bits: {
+    red: 100,
+    yellow: 128,
+    green: Number.MAX_SAFE_INTEGER,
+  },
+  max_bounty_amount: {
+    red: 64000, // $64k
+    yellow: 1000000, // $1M
+    green: Number.MAX_SAFE_INTEGER,
+  },
+} as const satisfies Record<string, MetricThresholds>
